Shrink font_slot_t fields to u16 where bounded
Replace u32 fields (atlas_tenant, sdf_range, atlas_w, atlas_h, ext_count) with u16 in font_slot_t, using bounds already enforced at load time (font_header_read and ORB_FONT_PAGE_MAX_W_SDF/ORB_FONT_MAX_GLYPHS). Adds a validation check in font_header_read to reject sdf_range values that exceed the page-width bound. Reorders struct fields to group small types together and reduce padding.

+++ source/runtime_service/gui/draw/gui_glyph_internal.c
@@ -78,7 +78,7 @@ font_slot_upload( font_slot_t* slot )
         }
     }
 
-    slot->atlas_tenant  = tenant;
+    slot->atlas_tenant  = (u16)tenant;     /* bounded by GUI_RES_ATLAS_MAX_TENANTS (320) */
     slot->tenant_sdf    = slot->sdf_range != 0;
     slot->needs_upload  = false;
     slot->upload_failed = false;
